third_party/WebKit/LayoutTests/fast/css/fontfaceset-check-platform-fonts.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
description('Tests FontFaceSet#check() returns true for platform fonts');

shouldBeTrue("document.fonts.check('10px Arial');");
shouldBeFalse("document.fonts.check('10px Nonexistent');");
shouldBeTrue("document.fonts.check('10px sans-serif');");
shouldBeTrue("document.fonts.check('10px Nonexistent, monospace');");
shouldBeFalse("document.fonts.check('10px Nonexistent1, Nonexistent2');");
</script>
